feat: 게시물 수정화면에서 지우는 파일 없는 경우도 가능

+++ backend/src/main/java/com/example/backend/service/board/BoardService.java
@@ -116,16 +116,18 @@ public class BoardService {
     }
 
     public boolean update(Board board, List<String> removeFiles) {
-        for (String file : removeFiles) {
-            String key = STR."prj1114/boardFile\{board.getBoardId()}/\{file}";
-            DeleteObjectRequest dor = DeleteObjectRequest.builder()
-                    .bucket(bucketName)
-                    .key(key)
-                    .build();
-            // s3 파일 지우기
-            s3.deleteObject(dor);
-            // db 파일 지우기
-            mapper.deleteFileByBoardIdAndName(board.getBoardId(), file);
+        if (removeFiles != null) {
+            for (String file : removeFiles) {
+                String key = STR."prj1114/boardFile\{board.getBoardId()}/\{file}";
+                DeleteObjectRequest dor = DeleteObjectRequest.builder()
+                        .bucket(bucketName)
+                        .key(key)
+                        .build();
+                // s3 파일 지우기
+                s3.deleteObject(dor);
+                // db 파일 지우기
+                mapper.deleteFileByBoardIdAndName(board.getBoardId(), file);
+            }
         }
 
         int cnt = mapper.update(board);
